==> 2_data_analysis/02_CRISPR_gene_dependency_scores_processing.ipynb <==
## CRISPR gene dependency scores  *-- Standard Z-score, Robust Z-score, Ranking*


In our project, we employ three approaches to identify SL interaction. The main idea is to compare the A1 gene dependecy scores. Here, we would mainly focus on the standard Z-score, robust Z-score, and the ranking of the gene dependecy scores.

**Input**
- BROAD CRISPR gene effect data: CRISPR_broad_paralog.csv

**Output**
- CRISPR gene dependecy scores: 
    - crispr_broad_paralog_standardZscore.csv
    - crispr_broad_paralog_robustZscore.csv
    - crispr_broad_paralog_ranking.csv

## Import modules
import numpy as np
import pandas as pd

## Load the dataset 
crispr_broad = pd.read_csv('/Users/amy/Desktop/SyntheticLethalityProject/1_data_processing/04_paralog_genes/crispr_broad_paralog.csv', index_col = None)
crispr_broad[:2]

**1. Standard Z-score**

$$\text{Standard }Z\text{-score} = \frac{{X_i – \mu}}{{\sigma}}$$

stand_z_gene_effect = crispr_broad.drop(['BROAD_ID', 'SangerModelID'], axis = 1).apply(lambda x: ((x-x.mean())/x.std(ddof=0))) 
stand_z_gene_effect = pd.concat([crispr_broad[['BROAD_ID', 'SangerModelID']], stand_z_gene_effect], axis = 1)
stand_z_gene_effect.to_csv('/Users/amy/Desktop/SyntheticLethalityProject/2_data_analysis/02_CRISPR_gene_dependency_scores_processing/crispr_broad_paralog_standardZscore.csv', index = False)
stand_z_gene_effect[:2]

**2. Robust Z-score**

$$\text{Robust }Z\text{-score} = \frac{{X_i – \tilde{X}}}{{MAD}}, \text{where }MAD = Median(|X_i - \tilde{X}|)$$

where

$$MAD = Median(|X_i - \tilde{X}|)$$

# calculate the robust z score of gene dependency value 
def robust_Zscore(x):
    
    '''Function for robust Z-score calculation'''
    
    median = np.median(x)
    MAD = np.median(np.abs(x-np.median(x)))
    robust_Zscore = (0.6745 * (x - median))/MAD
    
    return robust_Zscore
    
    
robust_z_gene_effect = crispr_broad.drop(['BROAD_ID', 'SangerModelID'], axis = 1).apply(robust_Zscore)
robust_z_gene_effect = pd.concat([crispr_broad[['BROAD_ID', 'SangerModelID']], robust_z_gene_effect], axis = 1)
robust_z_gene_effect.to_csv('/Users/amy/Desktop/SyntheticLethalityProject/2_data_analysis/02_CRISPR_gene_dependency_scores_processing/crispr_broad_paralog_robustZscore.csv', index = False)
robust_z_gene_effect[:2]

**3. Ranking**

rank_gene_effect = crispr_broad.drop(['BROAD_ID', 'SangerModelID'], axis = 1).apply(lambda x: x.rank(ascending=False))
rank_gene_effect = pd.concat([crispr_broad[['BROAD_ID', 'SangerModelID']], rank_gene_effect], axis = 1)
rank_gene_effect.to_csv('/Users/amy/Desktop/SyntheticLethalityProject/2_data_analysis/02_CRISPR_gene_dependency_scores_processing/crispr_broad_paralog_ranking.csv', index=False)
rank_gene_effect[:2]
